==> notebooks/05_Preprocess_final_dataset.ipynb <==
# Cleaning and Preprocessing the final dataset of publications related to the Tiny GenAI

# Importing the required libraries.
import csv, pandas as pd, numpy as np
from datetime import date

## 1. Generating the dataframe from the raw data

# Creating a dataframe from the raw data.
df_data = pd.read_csv("../data/raw/final_raw.csv", header=0, dtype=object)

# Checking the dataframe.
df_data.head()

# Visualizing the information of dataset.
df_data.info()

## 2. Cleaning and preprocessing the dataframe

# Function to normalize the affiliations of the authors.
def normalize_affiliations(row):
    # Getting missing values within the feature "author_affil" from "affiliations" one.
    if row.affiliations and row.author_affil:
        for pos, author in enumerate(row.author_affil):
            for affil in row.affiliations:
                if affil["id"] and author["affil_id"] and affil["id"] in [af.strip()
                        for af in author["affil_id"].split(",")]:
                    row.author_affil[pos]["affil_id"] = affil["id"]
                    row.author_affil[pos]["affiliation"] = affil["affiliation"]
                    if affil["country"] and not author["country"]:
                        row.author_affil[pos]["country"] = affil["country"]
                    elif affil["country"] != author["country"]:
                        row.author_affil[pos]["country"] = affil["country"]
    else:
        # Getting missing values within the feature "affiliations" from "author_affil" one.
        if row.author_affil:
            affils = set([(author["affil_id"], author["affiliation"], author["country"])
                          for author in row.author_affil
                          if author["affil_id"] or author["affiliation"]])
            if len(affils) > 0:
                keys = ["id", "affiliation", "country"]
                row.affiliations = tuple([dict(zip(keys, affil)) for affil in affils])
            else:
                row.affiliations = None
    return row

# Function to normalize the name of the authors.
def normalize_name_authors(row):
    if row.authors and row.author_affil:
        for pos, item in enumerate(row.authors):
            for author in list(row.author_affil):
                if item["id"] == author["id"]:
                    row.authors[pos]["name"] = author["name"]
    elif row.author_affil:
        authors = set([(author["id"], author["name"]) for author in row.author_affil
                       if author["name"]])
        if len(authors) > 0:
            keys = ["id", "name"]
            row.authors = tuple([dict(zip(keys, author)) for author in authors])
        else:
            row.authors = None

    return row

# Function to normalize the the authors and their affiliations.
def normalize_features(row):
    fields = {
        "authors": ["id", "name"],
        "affiliations": ["id", "affiliation", "country"],
        "affil": ["affil_id", "affiliation", "country"]
    }
    # Normalizing the authors.
    records = [tuple([item[f] for f in fields["authors"]]) for item in row.authors] \
        if row.authors else []
    if row.author_affil:
        records = set([*records, *[tuple([item[c] for c in fields["authors"]])
                                          for item in row.author_affil
                                          if item["id"] and item["name"]]])
    elif len(records) > 0 and not row.author_affil:
        row.author_affil = tuple([{**dict(zip(fields["authors"], auth)), "affil_id": None,
                                   "affiliation": None, "country": None} for auth in records])

    if len(records) > 0:
        row.authors = tuple([dict(zip(fields["authors"], auth)) for auth in records])

    # Normalizing the affiliations.
    if row.affiliations:
        records = [tuple([item[c] for c in fields["affiliations"]])
                          for item in row.affiliations]
        if row.author_affil:
            records = set([*records, *[tuple([item[c] for c in fields["affil"]])
                                              for item in row.author_affil
                                              if item["affil_id"] or item["affiliation"]]])
        row.affiliations = tuple([dict(zip(fields["affiliations"], affil))
                                  for affil in records])
    return row

# Defining the "None" value for the "NaN" values.
df_data.replace({np.nan: None}, inplace=True)

# Changing the type of features.
df_data.loc[:, ["auth_keywords", "index_terms", "affiliations", "subject_areas", "authors", "references"]] = \
    df_data.loc[:, ["auth_keywords", "index_terms", "affiliations", "subject_areas", "authors", "references"]].apply(
        lambda x: x.apply(lambda y: eval(y) if y else None), axis=1)
df_data.loc[df_data.data_source == "Scopus", "author_affil"] = df_data.loc[
    df_data.data_source == "Scopus", "author_affil"].apply(lambda x: eval(x) if x else None)
df_data.publication_date = pd.to_datetime(df_data.publication_date)
df_data.year = df_data.year.astype(np.int32)

# Fixing the value of the features "production_type" and "source_type".
df_data.loc[df_data.data_source == "Web of Science", ["production_type", "source_type"]] = \
df_data.loc[df_data.data_source == "Web of Science", ["production_type", "source_type"]].apply(
    lambda x: pd.Series({"production_type": "Journal", "source_type": "j"}) if "Article" in x.production_type \
        else pd.Series({"production_type": x.production_type, "source_type": x.source_type}), axis=1)

# Fixing the value of feature "language".
df_data.loc[df_data.data_source == "Web of Science", "language"] = df_data.loc[
    df_data.data_source == "Web of Science", "language"].apply(lambda x: x.lower()[:3])

# Defining the "zero" value for the articles without numbers of citation and references.
df_data.loc[df_data.citation_num.isnull(), "citation_num"] = 0
df_data.loc[df_data.ref_count.isnull(), "ref_count"] = 0

# Applying the function "normalize_name_authors" to the data.
df_data.loc[df_data.data_source == "Scopus", ["authors", "author_affil"]] = df_data.loc[
    df_data.data_source == "Scopus", ["authors", "author_affil"]].apply(
        normalize_name_authors, axis=1)

# Checking if there are the empty lists of authors.
df_data.authors[df_data.authors == ()].size

# Applying the function "normalize_affiliations" to the data.
df_data.loc[df_data.data_source == "Scopus", ["affiliations", "author_affil"]] = df_data.loc[
    df_data.data_source == "Scopus", ["affiliations", "author_affil"]].apply(
        normalize_affiliations, axis=1)

# Checking if there are the empty lists of affiliations.
df_data.affiliations[df_data.affiliations == ()].size

# Applying the function "normalize_features" to the data.
df_data.loc[df_data.data_source == "Scopus", ["authors", "affiliations", "author_affil"]] = df_data.loc[
    df_data.data_source == "Scopus", ["authors", "affiliations", "author_affil"]].apply(
        normalize_features, axis=1)

# Removing the unnecessary columns.
cols = ["language", "author_affil"]
df_data.drop(columns=cols, inplace=True)

# Checking the distribution of the feature "production_type".
pd.DataFrame(zip(df_data.production_type.value_counts(dropna=False).index,
                 df_data.source_type.value_counts(dropna=False).index,
                 df_data.production_type.value_counts(dropna=False).values
                 ), columns=["Production Type", "Source Type", "Quantity"]
)

# Applying the exclusion criteria.
df_data = df_data[df_data.year >= 2019]
df_data = df_data[df_data.abstract.notnull()]

# Defining the "None" value for the "NaN" values.
df_data.replace({np.nan: None}, inplace=True)

# Checking the dataframe.
df_data.head()

# Visualizing the information of dataset.
df_data.info()

## 3. Saving the dataframe

# Exporting the data to CSV file.
df_data.to_csv("../data/prepared/final_tiny_genai.csv", index=False, quoting=csv.QUOTE_ALL)
